#23 rename components to Pascal style - part1 (case changes are not detected by git)
